Update our_functs.h
Modification of word finding algorithm inside word_search function. Should work correctly now. If you want capitalization to matter, simply respond with 'y' or 'Y'. If not, respond with 'n' or 'N', and your inputted word, along with all words in text will be lowercased! Note the heavy placement of the cookie_monster() function. The monster has to constantly eat up the stdin cookies! Make a function in the future that will call scanf or getchar, and then call cookie_monster() from it, that way you have cleaner code.